+++ tests/testthat/test-matrix64.R
@@ -247,6 +247,7 @@ test_that("matrix multiplication", {
   expect_identical(mCo%*%t(m64), matrix(as.complex(m32%*%t(m32)), nrow=nrow(m32)))
 
   expect_error(m64%*%LETTERS[1:5], "non-numeric argument to binary operator", fixed=TRUE)
+  expect_error(m64%*%as.factor(LETTERS[1:5]), "non-numeric argument to binary operator", fixed=TRUE)
   expect_error(m64%*%as.raw(1:5), "non-numeric argument to binary operator", fixed=TRUE)
   
   # warning in multiplication part